requirements.txt
# Research Accountability & Mastery Platform - Phase 1A
# Python 3.11+ required

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Database
sqlalchemy[asyncio]>=2.0.27
asyncpg>=0.29.0
aiosqlite>=0.19.0
alembic>=1.13.1
greenlet>=3.0.0

# Authentication & Security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0

# Data Validation
pydantic>=2.6.0
pydantic-settings>=2.2.0
email-validator>=2.1.0

# Document Export
python-docx>=1.1.0

# Email
aiosmtplib>=3.0.0

# Utilities
python-dateutil>=2.8.2
diff-match-patch>=20230430

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Development
python-dotenv>=1.0.0
